requirements.txt
streamlit
pandas 
numpy
psycopg2
python-dotenv
